feat(global-amp-mode): configurable global default ampMode
BREAKING CHANGE: module does not provide middleware anymore
