+++ lib/module.js
@@ -16,6 +16,7 @@ module.exports = function (moduleOptions) {
   const options = {
     cdnBase: undefined,
     origin: '',
+    mode: 'hybrid',
     ...this.options.amp,
     ...moduleOptions
   }
